cdk/requirements.txt
-e .
aws-cdk.assets==1.25.0
aws-cdk.aws-cloudwatch==1.25.0
aws-cdk.aws-ecs==1.26.0
aws-cdk.aws-codebuild==1.26.0
